refactor: put logic in hook
